requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
openai==1.54.5
pyyaml==6.0.2
python-multipart==0.0.12